feat: load shared nav and modularize mobile menu
