+++ blog.html
@@ -11,32 +11,7 @@
   <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.3/font/bootstrap-icons.min.css">
 </head>
 <body id="body">
-  <header id="header" class="scrolled">
-    <div class="container header-container">
-      <a href="index.html" class="logo-link">
-        <img src="https://radilovk.github.io/bodybest/img/logoindex.png" alt="MyBody.Best" class="logo" />
-      </a>
-      <nav id="nav">
-        <ul>
-          <li><a href="index.html#hero" class="nav-link">Начало</a></li>
-          <li><a href="index.html#process" class="nav-link">Процесът</a></li>
-          <li><a href="index.html#tech" class="nav-link">Технология</a></li>
-          <li><a href="index.html#features" class="nav-link">Предимства</a></li>
-          <li><a href="index.html#programs" class="nav-link">Програми</a></li>
-          <li><a id="menu-personalization-link" href="personalization.html" class="nav-link">Персонализация</a></li>
-          <li><a href="#" class="nav-link open-modal-btn login-btn">Вход</a></li>
-          <li id="theme-toggle-li">
-            <button id="theme-toggle" aria-label="Смени темата">
-              <i class="bi bi-moon-stars-fill"></i>
-            </button>
-          </li>
-        </ul>
-      </nav>
-      <button class="mobile-menu-btn" aria-label="Toggle Menu">
-        <i class="fas fa-bars"></i>
-      </button>
-    </div>
-  </header>
+  <div id="nav-placeholder"></div>
 
   <main>
     <section class="container" style="padding-top:120px; padding-bottom:60px;">
