+++ index.html
@@ -18,34 +18,7 @@
     <link rel="stylesheet" href="style.css">
 </head>
 <body id="body">
-
-    <!-- Хедър - Непроменен -->
-    <header id="header" class="scrolled">
-        <div class="container header-container">
-            <a href="#hero" class="logo-link">
-                <img src="https://radilovk.github.io/bodybest/img/logoindex.png" alt="MyBody.Best" class="logo" />
-            </a>
-            <nav id="nav">
-                <ul>
-                    <li><a href="#hero" class="nav-link">Начало</a></li>
-                    <li><a href="#process" class="nav-link">Процесът</a></li>
-                    <li><a href="#tech" class="nav-link">Технология</a></li>
-                    <li><a href="#features" class="nav-link">Предимства</a></li>
-                    <li><a href="#programs" class="nav-link">Програми</a></li>
-                    <li><a id="menu-personalization-link" href="personalization.html" class="nav-link">Персонализация</a></li>
-                    <li><a href="#" class="nav-link open-modal-btn login-btn">Вход</a></li>
-                    <li id="theme-toggle-li">
-                        <button id="theme-toggle" aria-label="Смени темата">
-                            <i class="bi bi-moon-stars-fill"></i>
-                        </button>
-                    </li>
-                </ul>
-            </nav>
-            <button class="mobile-menu-btn" aria-label="Toggle Menu">
-                <i class="fas fa-bars"></i>
-            </button>
-        </div>
-    </header>
+    <div id="nav-placeholder"></div>
 
     <!-- Основно съдържание - Непроменено, с изключение на бутоните -->
     <main>
