+++ script.js
@@ -1,10 +1,10 @@
 import { toggleTheme, initializeTheme } from './js/themeControls.js';
+import { loadNav } from './js/nav.js';
+
+await loadNav();
 
 document.addEventListener('DOMContentLoaded', () => {
     const header = document.getElementById('header');
-    const body = document.body;
-    const nav = document.getElementById('nav');
-    const mobileMenuBtn = document.querySelector('.mobile-menu-btn');
     const themeToggleBtn = document.getElementById('theme-toggle');
     const openModalBtns = document.querySelectorAll('.open-modal-btn');
     let authModule;
@@ -18,37 +18,6 @@ document.addEventListener('DOMContentLoaded', () => {
             openAuthModal();
         });
     });
-
-    // Мобилно меню
-    if (mobileMenuBtn && nav) {
-        const closeNav = () => {
-            body.classList.remove('nav-open');
-            mobileMenuBtn.setAttribute('aria-expanded', 'false');
-        };
-        const toggleNav = () => {
-            const open = body.classList.toggle('nav-open');
-            mobileMenuBtn.setAttribute('aria-expanded', open);
-        };
-        mobileMenuBtn.addEventListener('click', toggleNav);
-        document.addEventListener('click', (e) => {
-            if (
-                body.classList.contains('nav-open') &&
-                !nav.contains(e.target) &&
-                !mobileMenuBtn.contains(e.target)
-            ) {
-                closeNav();
-            }
-        });
-        nav.addEventListener('click', (e) => { if (e.target === nav) closeNav(); });
-        nav.querySelectorAll('.nav-link').forEach(link => {
-            link.addEventListener('click', closeNav);
-        });
-    } else {
-        console.warn('script.js: липсва елемент за мобилно меню', {
-            btnExists: !!mobileMenuBtn,
-            navExists: !!nav
-        });
-    }
 
     // Плавно превъртане
     document.querySelectorAll('a[href^="#"]').forEach(anchor => {
